Various fixes

+++ packages/emacs/install.sh
@@ -42,7 +42,7 @@ doWork() {
 	ubuntu)
 	    # Override for Snap
 	    for d in "${cdir}"/desktop/*; do
-            cp -- "${d}" "${ddir}/emacs_$(basename ${d})"
+		cp -- "${d}" "${ddir}/emacs_$(basename ${d})"
 	    done
 	    ;;
 	*)
@@ -59,14 +59,18 @@ doWork() {
     mkdir -p ${edir}
     ln -sf "${cdir}"/environment.d/* ${edir}
 
-    local sf=/tmp/emacs-src.tar.gz
-    local sdir=~/.cache/emacs/c-src
-    curl -l -L https://github.com/emacs-mirror/emacs/archive/refs/tags/emacs-28.1.tar.gz -o ${sf}
-    rm -rf ${sdir}
-    mkdir -p ${sdir}
-    tar -xf ${sf} -C ${sdir}
-    mv ${sdir}/emacs-emacs-* ${sdir}/emacs
-    rm -f ${sf}
+    local version
+    version=$(emacs --version | grep -m 1 -Po 'GNU Emacs \K[.\d]+')
+    if [ -n "${version}" ]; then
+	local sf=/tmp/emacs-src.tar.gz
+	local sdir=~/.cache/emacs/c-src
+	curl -l -L https://github.com/emacs-mirror/emacs/archive/refs/tags/emacs-"${version}".tar.gz -o ${sf}
+	rm -rf ${sdir}
+	mkdir -p ${sdir}
+	tar -xf ${sf} -C ${sdir}
+	mv ${sdir}/emacs-emacs-* ${sdir}/emacs
+	rm -f ${sf}
+    fi
 
     cp "${cdir}"/emacs.service ${sysd}
     ln -sf ${sysd}/emacs.service ${sysdl}/emacs.service
